Create module for pkg/util/hostname/data
