+++ pkg/util/hostname/common.go
@@ -17,6 +17,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders/gce"
 	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 	"github.com/DataDog/datadog-agent/pkg/util/fargate"
+	"github.com/DataDog/datadog-agent/pkg/util/hostname/data"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname/validate"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -34,10 +35,7 @@ var (
 )
 
 // Data contains hostname and the hostname provider
-type Data struct {
-	Hostname string
-	Provider string
-}
+type Data = data.Data
 
 func fromConfig(ctx context.Context, _ string) (string, error) {
 	configName := config.Datadog.GetString("hostname")
